Refactor languageHandler to avoid import cycles
Also move its tests to vitest
