+++ apps/web/src/components/views/auth/CountryDropdown.tsx
@@ -10,7 +10,7 @@ import React, { type ReactElement } from "react";
 
 import { COUNTRIES, getEmojiFlag, type PhoneNumberCountryDefinition } from "../../../phonenumber";
 import SdkConfig from "../../../SdkConfig";
-import { _t, getUserLanguage } from "../../../languageHandler";
+import { _t, getCurrentLanguage } from "../../../languageHandler";
 import Dropdown from "../elements/Dropdown";
 import { type NonEmptyArray } from "../../../@types/common";
 
@@ -51,7 +51,7 @@ export default class CountryDropdown extends React.Component<IProps, IState> {
     public constructor(props: IProps) {
         super(props);
 
-        const displayNames = new Intl.DisplayNames([getUserLanguage()], { type: "region" });
+        const displayNames = new Intl.DisplayNames([getCurrentLanguage()], { type: "region" });
 
         this.countries = COUNTRIES.map((c) => ({
             name: displayNames.of(c.iso2) ?? c.iso2,
