+++ apps/web/src/components/views/elements/LanguageDropdown.tsx
@@ -10,13 +10,13 @@ Please see LICENSE files in the repository root for full details.
 import React, { type ReactElement } from "react";
 import classNames from "classnames";
 
-import * as languageHandler from "../../../languageHandler";
-import { _t } from "../../../languageHandler";
+import { _t, getCurrentLanguage } from "../../../languageHandler";
+import { getAllLanguagesWithLabels } from "../../../i18n/utils";
 import Spinner from "./Spinner";
 import Dropdown from "./Dropdown";
 import { type NonEmptyArray } from "../../../@types/common";
 
-type Languages = Awaited<ReturnType<typeof languageHandler.getAllLanguagesWithLabels>>;
+type Languages = Awaited<ReturnType<typeof getAllLanguagesWithLabels>>;
 
 function languageMatchesSearchQuery(query: string, language: Languages[0]): boolean {
     if (language.labelInTargetLanguage.toUpperCase().includes(query.toUpperCase())) return true;
@@ -48,8 +48,7 @@ export default class LanguageDropdown extends React.Component<IProps, IState> {
     }
 
     public componentDidMount(): void {
-        languageHandler
-            .getAllLanguagesWithLabels()
+        getAllLanguagesWithLabels()
             .then((langs) => {
                 langs.sort(function (a, b) {
                     if (a.labelInTargetLanguage < b.labelInTargetLanguage) return -1;
@@ -73,7 +72,7 @@ export default class LanguageDropdown extends React.Component<IProps, IState> {
         if (!this.props.value) {
             // If no value is given, we start with the first country selected,
             // but our parent component doesn't know this, therefore we do this.
-            const language = languageHandler.getUserLanguage();
+            const language = getCurrentLanguage();
             this.props.onOptionChange(language);
         }
     }
@@ -89,7 +88,7 @@ export default class LanguageDropdown extends React.Component<IProps, IState> {
             return <Spinner />;
         }
 
-        let displayedLanguages: Awaited<ReturnType<typeof languageHandler.getAllLanguagesWithLabels>>;
+        let displayedLanguages: Awaited<ReturnType<typeof getAllLanguagesWithLabels>>;
         if (this.state.searchQuery) {
             displayedLanguages = this.state.langs.filter((lang) => {
                 return languageMatchesSearchQuery(this.state.searchQuery, lang);
