+++ apps/web/src/test/setupGlobals.ts
@@ -1,0 +1,20 @@
+/*
+Copyright 2026 Element Creations Ltd.
+
+SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
+*/
+
+import { vi } from "vitest";
+
+import { mocks } from "../../test/setup/mocks.ts";
+
+// set up AudioContext API mock
+vi.stubGlobal("AudioContext", function () {
+    return mocks.AudioContext;
+});
+
+if (globalThis.window === undefined) {
+    // We are in a node environment, stub a basic window so singletons work
+    vi.stubGlobal("window", {});
+}
