+++ apps/web/src/test/setupLanguage.ts
@@ -6,20 +6,17 @@ SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Com
 Please see LICENSE files in the repository root for full details.
 */
 
-import fetchMock from "@fetch-mock/jest";
-import { ModuleLoader } from "@element-hq/element-web-module-api";
+import fetchMock from "@fetch-mock/vitest";
 import { merge } from "lodash";
-import { setMissingEntryGenerator } from "@element-hq/web-shared-components";
+import { setMissingEntryGenerator, setLanguage } from "@element-hq/web-shared-components";
 
-import { setLanguage } from "../../src/i18n/settings";
-import enElementWeb from "../../src/i18n/strings/en_EN.json";
-import deElementWeb from "../../src/i18n/strings/de_DE.json";
+import enElementWeb from "../i18n/strings/en_EN.json";
+import deElementWeb from "../i18n/strings/de_DE.json";
 // Cheat and import relatively here as these aren't exported by the module (should they be?)
 // eslint-disable-next-line no-restricted-imports
 import enSharedComponents from "../../../../packages/shared-components/src/i18n/strings/en_EN.json";
 // eslint-disable-next-line no-restricted-imports
 import deSharedComponents from "../../../../packages/shared-components/src/i18n/strings/de_DE.json";
-import { ModuleApi } from "../../src/modules/Api";
 
 const lv = {
     Save: "Saglabāt",
@@ -45,7 +42,6 @@ export function setupLanguageMock() {
     const enTranslations = merge(enElementWeb, enSharedComponents);
     const deTranslations = merge(deElementWeb, deSharedComponents);
 
-    fetchMock.mockGlobal();
     fetchMock
         .get(
             "end:/i18n/languages.json",
@@ -60,14 +56,11 @@ export function setupLanguageMock() {
         .get("end:de_DE.json", deTranslations)
         .get("end:lv.json", lv);
 }
-beforeEach(setupLanguageMock);
-afterEach(() => fetchMock.callHistory.flush());
 
 // Initialise the fetchMock before the test starts so the languageHandler.setLanguage call below can function
+fetchMock.mockGlobal();
+fetchMock.catch(404);
 setupLanguageMock();
+
 setLanguage("en");
 setMissingEntryGenerator((key) => key.split("|", 2)[1]);
-
-// Set up the module API (so the i18n API exists)
-const moduleLoader = new ModuleLoader(ModuleApi.instance);
-window.mxModuleLoader = moduleLoader;
