+++ apps/web/src/test/setupTests.ts
@@ -5,11 +5,12 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
 Please see LICENSE files in the repository root for full details.
 */
 
-import { vi, beforeEach } from "vitest";
+import { beforeEach, afterEach } from "vitest";
 import fetchMock, { manageFetchMockGlobally } from "@fetch-mock/vitest";
 
-import { mocks } from "../../test/setup/mocks.ts";
 import SdkConfig, { DEFAULTS } from "../SdkConfig";
+import "./setupGlobals.ts";
+import { setupLanguageMock } from "./setupLanguage.ts";
 
 manageFetchMockGlobally();
 
@@ -18,17 +19,11 @@ beforeEach(() => {
     fetchMock.hardReset();
     fetchMock.catch(404);
     fetchMock.mockGlobal();
+
+    setupLanguageMock();
 });
 
-// set up AudioContext API mock
-vi.stubGlobal("AudioContext", function () {
-    return mocks.AudioContext;
-});
-
-if (globalThis.window === undefined) {
-    // We are in a node environment, stub a basic window so singletons work
-    vi.stubGlobal("window", {});
-}
+afterEach(() => fetchMock.callHistory.flush());
 
 // uninitialised SdkConfig causes lots of warnings in console, init with defaults
 SdkConfig.put(DEFAULTS);
