+++ apps/web/src/utils/FormattingUtils.ts
@@ -10,7 +10,7 @@ Please see LICENSE files in the repository root for full details.
 import { type ReactElement, type ReactNode } from "react";
 import { useIdColorHash } from "@vector-im/compound-web";
 
-import { _t, getCurrentLanguage, getUserLanguage } from "../languageHandler";
+import { _t, getCurrentLanguage } from "../languageHandler";
 import { jsxJoin } from "./ReactUtils";
 
 export { formatBytes } from "@element-hq/web-shared-components";
@@ -69,7 +69,7 @@ export function formatList(items: ReactNode[], itemLimit = items.length, include
         return items[0] ?? "";
     }
 
-    const formatter = new Intl.ListFormat(getUserLanguage(), { style: "long", type: "conjunction" });
+    const formatter = new Intl.ListFormat(getCurrentLanguage(), { style: "long", type: "conjunction" });
     if (remaining > 0) {
         if (includeCount) {
             itemLimit--;
