+++ apps/web/test/unit-tests/components/views/elements/EventListSummary-test.tsx
@@ -130,7 +130,7 @@ describe("EventListSummary", function () {
 
     beforeEach(function () {
         jest.clearAllMocks();
-        jest.spyOn(languageHandler, "getUserLanguage").mockReturnValue("en-GB");
+        jest.spyOn(languageHandler, "getCurrentLanguage").mockReturnValue("en-GB");
     });
 
     afterAll(() => {
