+++ apps/web/test/unit-tests/components/views/messages/MPollBody-test.tsx
@@ -55,7 +55,7 @@ describe("MPollBody", () => {
 
         mockClient.getRoom.mockReturnValue(null);
         mockClient.relations.mockResolvedValue({ events: [] });
-        jest.spyOn(languageHandler, "getUserLanguage").mockReturnValue("en-GB");
+        jest.spyOn(languageHandler, "getCurrentLanguage").mockReturnValue("en-GB");
     });
 
     it("finds no votes if there are none", () => {
